+++ client/jackson/src/main/kotlin/net/corda/client/jackson/JacksonSupport.kt
@@ -17,7 +17,9 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.databind.SerializationFeature
 import com.fasterxml.jackson.databind.SerializerProvider
 import com.fasterxml.jackson.databind.deser.std.NumberDeserializers
+import com.fasterxml.jackson.databind.deser.std.StdDeserializer
 import com.fasterxml.jackson.databind.module.SimpleModule
+import com.fasterxml.jackson.databind.ser.std.StdSerializer
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
 import com.fasterxml.jackson.module.kotlin.KotlinModule
 import net.corda.core.contracts.Amount
@@ -27,7 +29,9 @@ import net.corda.core.crypto.AddressFormatException
 import net.corda.core.crypto.Base58
 import net.corda.core.crypto.Crypto
 import net.corda.core.crypto.MerkleTree
+import net.corda.core.crypto.PartialMerkleTree
 import net.corda.core.crypto.SecureHash
+import net.corda.core.crypto.SignatureMetadata
 import net.corda.core.crypto.TransactionSignature
 import net.corda.core.crypto.toStringShort
 import net.corda.core.identity.AbstractParty
@@ -117,15 +121,13 @@ object JacksonSupport {
             addDeserializer(OpaqueBytes::class.java, OpaqueBytesDeserializer)
             addSerializer(OpaqueBytes::class.java, OpaqueBytesSerializer)
 
-            addDeserializer(TransactionSignature::class.java, TransactionSignatureSerDe.Deserializer)
-            addSerializer(TransactionSignature::class.java, TransactionSignatureSerDe.Serializer)
+            listOf(TransactionSignatureSerde, SignedTransactionSerde).forEach { serde -> serde.applyTo(this) }
 
             // For X.500 distinguished names
             addDeserializer(CordaX500Name::class.java, CordaX500NameDeserializer)
             addSerializer(CordaX500Name::class.java, CordaX500NameSerializer)
 
             // Mixins for transaction types to prevent some properties from being serialized
-            setMixInAnnotation(SignedTransaction::class.java, SignedTransactionMixin::class.java)
             setMixInAnnotation(WireTransaction::class.java, WireTransactionMixin::class.java)
         }
     }
@@ -169,9 +171,31 @@ object JacksonSupport {
         registerModule(KotlinModule())
     }
 
-    object TransactionSignatureSerDe {
+    private interface JsonSerde<TYPE> {
+        val type: Class<TYPE>
+        val serializer: JsonSerializer<TYPE>
+        val deserializer: JsonDeserializer<TYPE>
 
-        object Serializer: JsonSerializer<TransactionSignature>() {
+        fun applyTo(module: SimpleModule) {
+            with(module) {
+                addSerializer(type, serializer)
+                addDeserializer(type, deserializer)
+            }
+        }
+    }
+
+    private inline fun <reified RESULT> JsonNode.get(fieldName: String, condition: (JsonNode) -> Boolean, mapper: ObjectMapper, parser: JsonParser): RESULT {
+
+        if (get(fieldName)?.let(condition) != true) {
+            JsonParseException(parser, "Missing required object field \"$fieldName\".")
+        }
+        return mapper.treeToValue(get(fieldName), RESULT::class.java)
+    }
+
+    object TransactionSignatureSerde : JsonSerde<TransactionSignature> {
+        override val type: Class<TransactionSignature> = TransactionSignature::class.java
+
+        override val serializer = object : StdSerializer<TransactionSignature>(type) {
             override fun serialize(value: TransactionSignature, json: JsonGenerator, serializers: SerializerProvider) {
                 with(json) {
                     writeStartObject()
@@ -184,11 +208,51 @@ object JacksonSupport {
             }
         }
 
-        object Deserializer: JsonDeserializer<TransactionSignature>() {
+        override val deserializer = object : StdDeserializer<TransactionSignature>(type) {
             override fun deserialize(parser: JsonParser, context: DeserializationContext): TransactionSignature {
-                TODO("sollecitom")
+                val mapper = parser.codec as ObjectMapper
+                val json = mapper.readTree<JsonNode>(parser)
+
+                if (json.get("by")?.isTextual != true) {
+                    JsonParseException(parser, "Missing required text field \"by\".")
+                }
+                val by = PublicKeyDeserializer.deserializeValue(json.get("by").textValue())
+                val signatureMetadata = json.get<SignatureMetadata>("signatureMetadata", JsonNode::isObject, mapper, parser)
+                val bytes = json.get<ByteArray>("bytes", JsonNode::isObject, mapper, parser)
+                val partialMerkleTree = json.get<PartialMerkleTree>("partialMerkleTree", JsonNode::isObject, mapper, parser)
+
+                return TransactionSignature(bytes, by, signatureMetadata, partialMerkleTree)
             }
         }
+    }
+
+    object SignedTransactionSerde : JsonSerde<SignedTransaction> {
+        override val type: Class<SignedTransaction> = SignedTransaction::class.java
+
+        override val serializer = object : StdSerializer<SignedTransaction>(type) {
+            override fun serialize(value: SignedTransaction, json: JsonGenerator, serializers: SerializerProvider) {
+                with(json) {
+                    writeStartObject()
+                    writeObjectField("txBits", value.txBits.bytes)
+                    writeObjectField("signatures", value.sigs)
+                    writeEndObject()
+                }
+            }
+        }
+
+        override val deserializer = object : StdDeserializer<SignedTransaction>(type) {
+            override fun deserialize(parser: JsonParser, context: DeserializationContext): SignedTransaction {
+                val mapper = parser.codec as ObjectMapper
+                val json = mapper.readTree<JsonNode>(parser)
+
+                val txBits = json.get<ByteArray>("txBits", JsonNode::isTextual, mapper, parser)
+                val signatures = json.get<TransactionSignatures>("signatures", JsonNode::isArray, mapper, parser)
+
+                return SignedTransaction(SerializedBytes(txBits), signatures)
+            }
+        }
+
+        private class TransactionSignatures : ArrayList<TransactionSignature>()
     }
 
     object ToStringSerializer : JsonSerializer<Any>() {
@@ -325,11 +389,15 @@ object JacksonSupport {
 
     object PublicKeyDeserializer : JsonDeserializer<PublicKey>() {
         override fun deserialize(parser: JsonParser, context: DeserializationContext): PublicKey {
+            return deserializeValue(parser.text, parser)
+        }
+
+        internal fun deserializeValue(value: String, parser: JsonParser? = null): PublicKey {
             return try {
-                val derBytes = parser.text.base64ToByteArray()
+                val derBytes = value.base64ToByteArray()
                 Crypto.decodePublicKey(derBytes)
             } catch (e: Exception) {
-                throw JsonParseException(parser, "Invalid public key ${parser.text}: ${e.message}")
+                throw JsonParseException(parser, "Invalid public key $value: ${e.message}")
             }
         }
     }
@@ -373,7 +441,7 @@ object JacksonSupport {
     }
 
     abstract class SignedTransactionMixin {
-        @JsonIgnore abstract fun getTxBits(): SerializedBytes<CoreTransaction>
+        @JsonProperty("txBits") protected abstract fun getTxBits(): SerializedBytes<CoreTransaction>
         @JsonProperty("signatures") protected abstract fun getSigs(): List<TransactionSignature>
         @JsonProperty protected abstract fun getTransaction(): CoreTransaction
         @JsonIgnore abstract fun getTx(): WireTransaction